Changes: Updated button background drawing to be more efficient.
RevBy: Pekka Vuorela
Details: Some keyboard layouts have so many keys in them that they
have to be shrinked in width. This patch applies a performance fix
made earlier also to cover these cases. The main idea is to prevent
DuiScalableImage from actually scaling the image in most cases, because
it's a bit faster that way.