Fixes: Fixes Ut_MKeyboardHost failures due to a bug related to orientation change.
RevBy: Hannu Koivisto
Details: MWindow and scenemanager can get out of sync and then it's not
possible to change orientation. While waiting for fix, let's just use
only one interface for changing orientation, choosing MWindow for now.
Also, another bug is that orientation finished is sent if window is visible
and 180 degree rotation is made. Changed the order of rotations to always
get the signal.