Fixes: NB#161456 "Dui-im-uiserver crashes while changing Languages"
RevBy: Hannu Koivisto, Ning Chi, Mohammad Anwari
Details: The index handling of the HorizontalSwitcher has been improved, and
MVirtualKeyboard now properly checks for invalid widgets returned by the
HorizontalSwitcher (since that is documented to happen in the
HorizontalSwitcher API).

Then, MVirtualKeyboard::languageReset() did not update the current language for
the failure case (found by Ning).

It also turned out that HorizontalSwitcher::removeWidget() was only used by a test, hence it got removed.
Fixed tests for HorizontalSwitcher::setCurrent() for the case when an widget
was added to a yet empty HorizontalSwitcher container.