Fixes: NB#166341, hwkbd autocaps, and state handling in general
RevBy: Chi Ning
Details: don't reset() hwkbd in MKeyboardHost::show() and hide() because they
might not even be called in the case of hardware keyboard.  They are also the
wrong time to reset() considering when host calls setAutoCapitalization().
It also turns out that reset() doesn't need to and shouldn't do anything in
case of hwkbd but instead state, which is mainly the external keyboard state,
should be handled based on focus in/out events for which there is now a new
interface offered by meego-im-framework.  Also enable autorepeat, disable
key redirection and unlatch & unlock everything on focus out.