Changes: Remove broken convenience API
RevBy: Jarno Malmari
Details: HorizontalSwitcher had 3 function overloads for switchTo(), all with a
very different implementations. Only switchTo(Direction) turned out to have the
correct implementation: switchTo() is meant to be used for changing the KBD
layout with a smooth transition. That only works with KBD layouts that are
right next to each other, hence the decision to remove the function overloads.