Fixes: update vkb position on layout changes
RevBy: Jarno Malmari, Viacheslav Sobolev, Michael Hasselmann
Details: if the original layout and the new layout had a different number of
rows, the vkb position was not updated and so, for example, you could no
longer see the function row.  Also update position on all toolbar region
changes.  This is an intermediate solution; a better design will most likely
involve changing organizeContent() not to do repositioning (instead do that
in a resize event handler) and removing actualHeight() (arrange things so
that size() can always be used directly).