Fixes: NB#179381, Display pressed keys using pre-edit
RevBy: Pekka Vuorela, Ning Chi
Details: also, events are no longer sent during release or when long press is
triggered; instead a commit string is sent.  Shift and Fn modifier events are
no longer passed to the application.  Symbol key features and long press were
modified to work in the spirit of this new "action on key down" scheme.  This
commit doesn't remove some of the trickery that was needed for "action on key
up" and which is now obsolete (but not harmful); removing that is considered
to be optimization and can be done separately, keeping this change as minimal
as possible in order to achieve the new functionality.