Changes: Remove vertical spacing used for adjacent KeyButtonAreas.
RevBy: Hannu Koivisto
Details: KeyButtonAreas are no longer piled up on each others and the
spacing is not used. This is done in preparation to put the spacing
attribute in KeyButtonArea specific style. If the removed spacing is
needed at some point, a new attribute should be introduced for that.