Fixes: Reaction map tests for MVirtualKeyboard
RevBy: Ning Chi
Details: The MImOverlay used a translucent but "visible" MSceneWindow. This
caused our reaction map test to fail like so:
1. Find this special scene item,
2. Decide that this item should not touch the reaction map,
3. Look at the actual pixel drawn on screen (by the underlying key area)
4. Get utterly confused and goto fail.

To avoid translucent-but-visible scene items (which MReactionMapTester cannot
handle), I removed the MSceneWindow parent of MImOverlay, and let MImOverlay
inherit from MSceneWindow instead of MWidget. That way, we can still use it
together with MSceneManager.

Also, the slot name that the reaction map tried to use was long renamed/changed
in the VKB, so this got fixed, too. And it did not like the bounding rect
adjustment stuff.