Changes: refactor toolbar and vkb show/hide logic
RevBy: Chi Ning
Details: a) vkb is shown only when in OnScreen state, b) when
there's nothing on screen (hwkb mode, empty toolbar), region is empty and
therefore the passthru window is not mapped, c) toolbar is shown/hidden based
on SIP open/close requests (in hwkb state too), d) we no longer connect
toolbar region update signal to shared handle area region updating in two
places.  As a consequence to a), vkb no longer fades the shared handle area
which is a loss wrt. transitions, but transitions need fixing for other
reasons as well, so at that point the desired transition should be added to
the shared handle area itself.