Fixes: vkb: emit layoutChanged after layout has been changed, not before
RevBy: Pekka Vuorela, Kimmo Surakka, Chi Ning
Details: if you called MVirtualKeyboard::setLayout(foo); and in the handler
of the emitted layoutChanged signal MVirtualKeyboard::mainLayoutKeys(), you
would have gotten keys for the old layout if setLayout() actually needed to
change the layout (which it usually doesn't need to do and it's possible that
when it does, mainLayoutKeys() result isn't actually used, but let's play it
safe).  This is a temporary fix, layout switching should be later changed so
that we can rely on HorizontalSwitcher's onSectionSwitched.