Fixes: NB#235747 [TASK] Backspace icon on virtual keyboard is too small
RevBy: Michael Hasselmann, Pekka Vuorela
Details: Automatically fall back to a smaller icon.

Two icons can be specified for every key now; the
normal icon and a "compact" icon. The compact icon will be
used if the normal icon does not fit in the key, as determined
by the required-key-icon-margins CSS attribute.