Fixes: NB#244195, Application with keyboard open in landscape mode can be iconified but immediately comes back
RevBy: Pekka Vuorela
Details: don't let orientation change code screw up region signal suppression
set up for hide animation.  When the application/kbd is iconified in
landscape, we get hide() and then the orientation changes to portrait.  This
caused region updates to be enabled and non-empty region to be sent.  The
non-empty region caused QWidget::showFullScreen() to be called for the
passthruwindow, because the widget was already invisible at that point.

One could argue that orientation change should not be notified when SIP has
not been requested.  That could be handled in MInputContext, but one might
also argue that when the application is hidden, there is no active window and
thus active window orientation doesn't change and MComponentData should not
call MInputMethodState orientation methods (but maybe that's not the way
active window is defined in LMT, haven't checked).