Fixes: Crashing UT's due to std::bad_cast in RegionStore
RevBy: Pekka Vuorela, Hannu Koivisto
Details: RegionStore is cleaned up on QObject::destroy, but combineRegion
depends on a dynamic_cast (in order to retrieve the widget geometry). Added the
missing check for the cast now, but that does not solve the problem of the
RegionStore storing regions from dead widgets.