Fixes: Crash in Ut_MImKey::initTestCase
RevBy: Pekka Vuorela
Details: When LMT is build in debug mode, MTheme::release will fatally abort if
asked to release a pixmap that was not requested through MTheme::pixmap. Since
the UT takes care of deleting the pixmaps, the fix is simply to fix up the
pointers in MImKey::IconInfo before MImKey gets destroyed.