Fixes: latched shift when extended key area is shown
RevBy: Michael Hasselmann
Details: don't reset active keys in ExtendedKeys::showExtendedArea().  Since
we're sending a cancel event to the main key area, the reset in ExtendedKeys
is duplicate work and suffers from problems recently fixed in the
MImAbstractKeyArea::reset() called by cancel event handling.  In particular,
one of the problems was that latched shift was reseted at the time of
showing the extended key area.