Changes: remove last active key reseting from key area's ungrab handler
RevBy: Michael Hasselmann
Details: if there can be a key pressed, we expect cancel event to be sent and
cancel handler resets correct active keys.  This last active key touch point
reseting was also simply wrong because it didn't check if the last active key
was this key area's key.