Fixes: inconsistency between shift key state and level of other keys
RevBy: Michael Hasselmann
Details: always send a cancel event when showing symbol view, even when we
are not opening a temporary symbol view (which happens on symbol key press
and cancel is obviously needed then) but also when symbol key is tapped,
because user may be holding shift down and sym is not the primary touchpoint.
Also in this case we want to reset the active key area so that when the
symbol view is exited, keys won't be in level 1 even when shift key state is
normal.