Fixes: NB#277034 Input Methods: The right and left brackets are improperly displayed in Arabic Virtual Keyboard
RevBy: Ning Chi
Details:
The Right-to-Left Embedding mark is prepended to the label before drawing on the key and magnifier.
The label which is used to send the text is still preserved. This approach works fine if the text
direction on the entry is already set to RTL (eg. when there's already Arabic text previously or
the text direction is explicitly set to RTL).