New: Add init method for MKeyboardHost
RevBy: TrustMe
Details: MKeybhoardHost can be accessed through a singleton, but during
construction, that singleton might not be available yet. This kind of limits
the places where and when the MKeybhoardHost singleton can be reliably used. To
workaround that, we keep the c'tor minimal and defer remaining initialization
to init method. Compared to the previous c'tor, we can now guarantee that the
code in init has access to a fully constructed MKeyboardHost instance (which is
still a bit different than just moving the singleton creation to the top of the
original c'tor). For example, invocation of init could now be deferred to the
mainloop.
For convenience, a static create method has been added as well.