Fixes: Use rpath to dynamically link Ut_WidgetBar to meego-keyboard lib
RevBy: Jon Nordby
Details: Setting a relative path explicitly  fails when the test is run after
being installed. Therefore, rpath is used.
To override rpath, use LD_LIBRARY_PATH - when using make check, that one is
automatically set to source directory containing libmeego-keyboard.so, allowing
the test to run against the compiled version of the lib, not the installed
version.
Also removed the unnecessarily included headers and source files, to speed up
compile time and to be on the safe side.