Remove active keys API from KeyArea and attach it to Layout instead
RevBy: TrustMe
Details: Realized that I was lying about the active keys; the renderer's
KeyAreaItem doesn't need it at all. Also saves me from having to solve the
active keys sync issue between copies of KeyArea's. We need the active keys for
the KeyItem overlays, but in that case the renderer has the Layout context
anyway.