Fix remaining Qt 5 build errors
RevBy: TrustMe
QtGui does not contain Qwidget/QApplication stuff any more, so we use
individual header includes instead. This is the only way to keep it compilable
for Qt 4 without falling back to ifdefs. Luckily, this annoyance is mostly
contained in libmaliit-keyboard-view.
The language layout loader benchmark had to be disabled for Qt 5, because it
still requires QApplication to run.