Fix parsing layouts using old-style imports.
RevBy: Michael Hasselmann, Jan Arne Petersen
LayoutParser expects no child tags when old-style import tag
is used. For this it used readNextStartElement() which was wrong
because it was reading a file to its end when import tag was last
opening tag in file. Replaced it with some more scrutinous check.