Fix crash in maliit-keyboard-viewer
RevBy: Jan Arne Petersen
Commit 3ab4090, "Use surface API in the plugin", does not set a surface factory
for the renderer when run as maliit-keyboard-viewer (that is, in stand-alone
mode, in contrast to Maliit plugin mode).  This patch checks against an invalid
main surface when trying to show the keyboard and prints a critical warning in
that case.

Also adds a missing return statement to the next section, as the code comment
seems to indicate that we should return here, too.