Do not autodisable null feedback.
RevBy: Michael Hasselmann

The "feedback-enabled" is user setting, so it is not propagated to
maliit-server at runtime and that can easily make a discrepancy
between setting in settings example and the actual enabled
value. Also, null feedback does not do anything, so there is really no
difference whether it is enabled or not.

This work was done by Openismus GmbH for Canonical Ltd