Fix rendering errors when changing bounding rects of graphics items
RevBy: Jan Arne Petersen, Jon Nordby
When implicitly changing the boundingRect() of a QGraphicsItem (which is what
we do when setting for instance a new KeyArea to a KeyAreaItem), we need to
call QGraphicsItem::prepareGeometryChange() beforehand. Otherwise, Qt
GraphicsView cannot properly update the scene graph.

This work was done by Openismus GmbH for Canonical Ltd