New: Add auto-caps after commiting punctuated preedit
RevBy: Jon Nordby
Let the abstract editor decide when to trigger auto-caps, but delegate the
layout update to the LayoutUpdater. For that, we simply need to connect the
editor's new autoCapsActivated signal to the existing autoCapsActivated signal
of the LayoutUpdater. Had to enforce layout sync in state machine for
latched-shift state though.

This work was done by Openismus GmbH for Canonical Ltd