Allow to save user dictionary for spellchecking
Hunspell does not directly support an user dictionary. We use a plain
UTF8 text file with a word per line, which is read on startup. When a
new word is added to the user dictionary it is appended to the user
dictionary file.

RevBy: Michael Hasselmann

This work was done by Openismus GmbH for Canonical Ltd