Remove "extended_labels" in favour of <extended> tag.
RevBy: Michael Hasselmann

"extended_labels" was an attribute storing string of labels of
extended keys. It was quite limited, because it does not allow us to
easily specify labels with multiple characters or extended keyboards
with more than one row or keys with icons. So instead of:

<binding label="e" extended_labels="ęéě"/>

we should use:

<extended>
  <row>
    <key>
      <binding label="ę"/>
    </key>
    <key>
      <binding label="é"/>
    </key>
    <key>
      <binding label="ě"/>
    </key>
  </row>
</extended>

This change makes specifying extended keys more verbose, but it gives
us a possilibity to specify more than one level of extended keyboards
(which might be needed for some languages), more than one row in
extended keyboard, icons in extended keys, extended keys with some
special action and so on.

This breaks current layout files we have, but those will be converted.

This work was done by Openismus GmbH for Canonical Ltd