Fix closing of extended panel by clicking outside of it
RevBy: TrustMe
When switching to surface API, we lost the ability to close the extended panel
by clicking outside of it. This probably went unnoticed because it *still*
worked for the special case where we would click on the main panel.

Notice how we use keyAreaReleased as a signal with cancel semantics now, in the
Glass. The [grab|release]Mouse code in Renderer needs to be checked with touch
events.

This work was done by Openismus GmbH for Canonical Ltd