Fixes: NEMO#451 - Use a single MouseArea for the whole keyboard
With a single MouseArea, the Flickable used for the swipe-down gesture
is no longer needed. This fixes bugs with the Flickable stealing press
events, and accidentally dragging the keyboard all over the screen.

Also, with a single MouseArea, keys are now registered only on release,
and it's possible to slide between keys before releasing, which improves
usability dramatically. That also fixes NEMO#451, because the swipe down
gesture cannot cause random keypresses.

RevBy: Pekka Vuorela, Robin Burchell