Make extended keys a separate surface with separate QML file
The extended (overlay) surface would need to be parented to the main surface,
but that currently does not show anything. Not parenting, however, shows the
extended surface at the correct time but in the wrong spot.

Would be nice to find out whether two surfaces can indeed share the same QML
engine, but that needs API changes in the framework.

RevBy: Jan Arne Petersen

This work was done by Openismus GmbH for Canonical Ltd