Maliit Keyboard: Fix shift behaviour
There was an extra shift state in the state machine when in fact there are only
3 states for shift behaviour: no shift, shift latched and caps locked.

The actual problem is that when we switch to shift view, we forget that the
shift key was pressed. Whenever that happens, we won't receive a
shiftReleased() signal. But that requires a bigger rewrite to be properly
fixed.

RevBy: TrustMe